requirements.txt
allure-behave==2.8.0
allure-python-commons==2.8.0
behave==1.2.6
playwright==1.48.0
pyyaml==6.0
